New datalad.support.extensions with register_config()
As requested in
https://github.com/datalad/datalad/pull/6601#issuecomment-1083122326
